Update README.md
I worked on this for 8 hours with Chris Champness and David Dyer, then I screwed up all the code, had to start over finished it, so this is my code. I am currently evacuating from New Orleans and Hurricane Ida.  
I wasn't sure if I needed to keep the other files, so you could see how well the website works. If this is not needed then I will not include it next time. Thank you, and I hope you have nice day. 
I am going to upload of the website now as a reference as suggested in my class. I hope you have a nice day.